Main app files
Main Java source files for grocery app and also folder containing xml files for the GUI. This does not include original folder structure and all original files for app,  meaning it will likely not run if it is cloned. Just wanted to show my main source code for building this app.